# app/static/swagger.yaml
...
swagger: "2.0"
...
info:
  title: "Mechanic API" # Title of the API
  description: "API to manage a Mechanic Shop" # A brief description of what the API does
  version: "1.0.0" # The version of your API

# The base URL of the API, in this case, "127.0.0.1:5000". This is where the API can be accessed.
host: "127.0.0.1:5000"
...
schemes:
  - "http"

# The types of request body formats the API can consume. Here, it's "application/json",
# meaning the API can accept JSON-formatted request bodies.
consumes:
  - "application/json"

# The types of response body formats the API can produce. Here, it's also "application/json",
# indicating that the API will respond with JSON-formatted data.
produces:
  - "application/json"

# Add securityDefinitions to allow endpoints to require jwt (i.e. tokens) for access
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header

# Paths - This is where you would define the endpoints of your API, along with the HTTP methods
# they support and their respective request/response formats.

# -------------------- Customer Endpoints --------------------
paths:
  /customers/login: # ⚡ Tested!
    post:
      tags:
        - Customers
      summary: "Login endpoint, (generates token)"
      description: "Validates customer credentials, generates and returns a customer specific token."
      parameters:
        - in: "body"
          name: "body"
          description: "Customer login credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Successful login"
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              token: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyX2lkIjoxLCJleHAiOjE2ODgwMDAwMDB9.XYZ123abc456DEF789ghi012JKL345mno678PQR90stu"
              message: "Login successful"
              status: "success"
  /customers: # ⚡ Tested!
    post:
      tags:
        - Customers
      summary: "Endpoint to create/register a new customer"
      description: "Endpoint to create a new customer in our Mechanic Shop system. Emails must be unique"
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a new customer"
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          description: "Customer created successfully"
          schema:
            $ref: "#/definitions/CreateCustomerResponse"
          examples:
            application/json:
              id: 1
              name: "Puka Roo"
              email: "pukaroo@example.com"
              phone: "123-456-7890"
              password: "securepassword123"

    get: # ⚡ Tested!
      tags:
        - Customers
      summary: "Returns all customers"
      description: "Endpoint to retrieve a list of customers in the Mechanic Shop system."
      responses:
        200:
          description: "Retrieved Members Successfully"
          schema:
            $ref: "#/definitions/AllCustomers"
  /customers/{customer_id}: # ⚡ Tested!
    get:
      tags:
        - Customers
      summary: "Get a specific customer by ID"
      description: "Retrieve a specific customer's details using their unique customer ID."
      parameters:
        - name: customer_id
          in: path
          required: true
          description: "ID of the customer to retrieve"
          type: integer
      responses:
        200:
          description: "Customer retrieved successfully"
          schema:
            $ref: "#/definitions/CreateCustomerResponse"
          examples:
            application/json:
              id: 1
              name: "Puka Roo"
              email: "pukaroo@example.com"
              phone: "123-456-7890"
        404:
          description: "Customer not found"

    put: # ⚡ Tested!
      tags:
        - Customers
      summary: "Endpoint to update a customer."
      description: "Endpoint to update a customer. This is a token authenticated route, and the customer must be logged in to update their account"
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to update a customer."
          required: true
          schema:
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          description: "Customer updated successfully"
          schema:
            $ref: "#/definitions/UpdateCustomerResponse"
          examples:
            application/json:
              id: 1
              name: "Puka Roo Updated"
              email: "pukaroo@example.com"
              phone: "123-456-7890"
              password: "securepassword123"

    delete: # ⚡ Tested!
      tags:
        - Customers
      summary: "Endpoint to delete a customer."
      description: "Endpoint to delete a customer. This is a token authenticated route, and the customer must be logged in to delete their account."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Customer deleted successfully"
          schema:
            $ref: "#/definitions/DeleteCustomerResponse"
          examples:
            application/json:
              message: "Customer <customer_id> deleted successfully."
              status: "success"

  # Search for a customer based on email # ⚡ Tested!
  /customers/search:
    get:
      tags:
        - Customers
      summary: "Search for a customer by email"
      description: "Retrieve a customer's details using their email address. Email must match exactly."
      parameters:
        - name: email
          in: query
          required: true
          description: "Email address of the customer to search for"
          type: string
      responses:
        200:
          description: "Customer retrieved successfully"
          schema:
            $ref: "#/definitions/CustomerListResponse"
          examples:
            application/json:
              id: 1
              name: "Puka Roo"
              email: "pukaroo@example.com"
              phone: "123-456-7890"
        404:
          description: "Customer not found"

  # -------------------- Inventory Endpoints --------------------
  /inventory: # ⚡ Tested!
    post:
      tags:
        - Inventory
      summary: "Endpoint to create/register a new inventory item"
      description: "Endpoint to create a new inventory item in our Mechanic Shop system. Part name must be unique"
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a new inventory item"
          required: true
          schema:
            $ref: "#/definitions/CreateInventoryPayload"
      responses:
        201:
          description: "Inventory item created successfully"
          schema:
            $ref: "#/definitions/CreateInventoryResponse"
          examples:
            application/json:
              id: 1
              part_name: "Caliper"
              price: "56.99"
              service_ticket_inventory: [1, 2, 3]

    get: # ⚡ Tested!
      tags:
        - Inventory
      summary: "Returns all inventory items"
      description: "Endpoint to retrieve a list of inventory items in the Mechanic Shop system."
      responses:
        200:
          description: "Retrieved Inventory Items Successfully"
          schema:
            $ref: "#/definitions/AllInventoryItems"
  /inventory/{id}: # ⚡ Tested!
    get:
      tags:
        - Inventory
      summary: "Get a specific inventory item by ID"
      description: "Retrieve a specific inventory item using the unique ID."
      parameters:
        - name: id
          in: path
          required: true
          description: "ID of the inventory item to retrieve"
          type: integer
      responses:
        200:
          description: "Inventory item retrieved successfully"
          schema:
            $ref: "#/definitions/CreateInventoryResponse"
          examples:
            application/json:
              id: 1
              part_name: "Caliper"
              price: "56.99"
              service_ticket_inventory: [1, 2, 3]
        404:
          description: "Inventory item not found"

    put: # ⚡ Tested!
      tags:
        - Inventory
      summary: "Endpoint to update an inventory item."
      description: "Endpoint to update an inventory item. This is a token authenticated route, and the mechanic must be logged in to update an item."
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: "ID of the inventory item to update"
          type: integer
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/UpdateInventoryPayload"
      responses:
        200:
          description: "Inventory item updated successfully"
          schema:
            $ref: "#/definitions/UpdateInventoryResponse"
          examples:
            application/json:
              id: 1
              part_name: "Caliper"
              price: "56.99"
              service_ticket_inventory: [1, 2, 3]

    delete: # ⚡ Tested!
      tags:
        - Inventory
      summary: "Endpoint to delete an inventory item."
      description: "Endpoint to delete an inventory item. This is a token authenticated route, and the mechanic must be logged in to delete an item."
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: "ID of the inventory item to delete"
          type: integer
      responses:
        200:
          description: "Inventory item deleted successfully"
          schema:
            $ref: "#/definitions/DeleteInventoryResponse"
          examples:
            application/json:
              message: "Inventory item <id> deleted successfully."
              status: "success"
  /inventory/search: # ⚡ Tested!
    get:
      tags:
        - Inventory
      summary: "Search for an inventory item by part name"
      description: "Retrieve an inventory item's details using the unique part name."
      parameters:
        - name: part_name
          in: query
          required: true
          description: "Part name of the inventory item to search for"
          type: string
      responses:
        200:
          description: "Inventory item retrieved successfully"
          schema:
            $ref: "#/definitions/CreateInventoryResponse"
          examples:
            application/json:
              id: 1
              part_name: "Caliper"
              price: "56.99"
              service_ticket_inventory: [1, 2, 3]
        404:
          description: "Inventory item not found"

  # -------------------- Mechanic Endpoints --------------------
...
definitions:
...
  LoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
      password:
        type: "string"

    required:
      - "email"
      - "password"

  # The object, named LoginResponse, is designed to represent the structure of the response that
  # the API will send when a customer makes a POST request to the /customers/login endpoint.

  # -------------------- Customer Definitions --------------------
  LoginResponse:
    type: "object"
    properties:
      token:
        type: "string"
      message:
        type: "string"
      status:
        type: "string"
  CreateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
    required:
      - name
      - email
      - phone
      - password
  CreateCustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
  AllCustomers:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        name:
          type: "string"
        email:
          type: "string"
        phone:
          type: "string"
        password:
          type: "string"
  UpdateCustomerPayload:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
    required:
      - name
      - email
      - phone
      - password
  UpdateCustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"
      email:
        type: "string"
      phone:
        type: "string"
      password:
        type: "string"
  DeleteCustomerResponse:
    type: "object"
    properties:
      message:
        type: "string"
      status:
        type: "string"
  CustomerListResponse:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
        name:
          type: "string"
        email:
          type: "string"
        phone:
          type: "string"
        password:
          type: "string"

  # -------------------- Inventory Definitions --------------------
  CreateInventoryPayload:
    type: "object"
    properties:
      part_name:
        type: "string"
      price:
        type: "number"
        format: "float"

    required:
      - part_name
      - price
  CreateInventoryResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      part_name:
        type: "string"
      price:
        type: "number"
        format: "float"
      service_ticket_inventory:
        type: "array"
        items:
          type: "integer"
  AllInventoryItems:
...
  UpdateInventoryPayload:
    type: "object"
    properties:
      part_name:
        type: "string"
      price:
        type: "number"
        format: "float"
    required:
      - part_name
      - price
  UpdateInventoryResponse:
    type: "object"
    properties:
      id:
        type: "integer"
      part_name:
        type: "string"
      price:
        type: "number"
        format: "float"
      service_ticket_inventory:
        type: "array"
        items:
          type: "integer"
  DeleteInventoryResponse:
    type: "object"
    properties:
      message:
        type: "string"
      status:
        type: "string"

  # -------------------- Mechanic Definitions --------------------
...